# public/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QuakeWise External API",
    "version": "1.0.0",
    "description": "Comprehensive earthquake safety assessment API combining AI analysis, geospatial data, and structural engineering calculations. Provides building vulnerability assessments for seismic risk evaluation.",
    "contact": {
      "name": "QuakeWise API Support",
      "email": "api@quakewise.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://quakewise.com/terms"
    }
  },
  "servers": [
    {
      "url": "https://quakewise.com/api/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "PlatformToken": [],
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/utils/proxy-image": {
      "get": {
        "tags": ["Utilities"],
        "summary": "Proxy Google Maps images",
        "description": "CORS proxy for Google Maps images including Street View and satellite imagery. Caches images for improved performance.",
        "operationId": "proxyImage",
        "security": [{"PlatformToken": [], "BearerAuth": []}],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Encoded Google Maps image URL"
          }
        ],
        "responses": {
          "200": {
            "description": "Image retrieved successfully",
            "content": {
              "image/jpeg": {"schema": {"type": "string", "format": "binary"}},
              "image/png": {"schema": {"type": "string", "format": "binary"}}
            }
          },
          "400": {"$ref": "#/components/responses/ValidationError"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "429": {"$ref": "#/components/responses/RateLimitExceeded"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "PlatformToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Platform-Token",
        "description": "Static platform authentication token provided by QuakeWise"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token issued via /auth/issue-token endpoint"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {"type": "string"},
              "message": {"type": "string"},
              "details": {"type": "object"},
              "timestamp": {"type": "string", "format": "date-time"},
              "requestId": {"type": "string"}
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication failed",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        },
        "headers": {
          "X-Error-Code": {"schema": {"type": "string"}}
        }
      },
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      },
      "RateLimitExceeded": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        },
        "headers": {
          "X-RateLimit-Limit": {"schema": {"type": "integer"}},
          "X-RateLimit-Remaining": {"schema": {"type": "integer"}},
          "X-RateLimit-Reset": {"schema": {"type": "string", "format": "date-time"}},
          "X-RateLimit-Tier": {"schema": {"type": "string"}}
        }
      },
...
    }
  }
}